requirements.txt
Pillow
moment
python-mpd2
pyLMS
redis
pyOWM
luma.oled
socketIO-client
